Extract getApiKey to separate file to fix circular dependency
