+++ packages/client/src/api/client.ts
@@ -1,4 +1,4 @@
-import { getAPIKey } from '../util/config';
+import { getAPIKey } from '../util/apiKey';
 import { getFetchImplementation } from '../util/fetch';
 import type * as Types from './components';
 import { operationsByTag } from './components';
