+++ packages/client/src/util/apiKey.ts
@@ -1,0 +1,9 @@
+import { getEnvVariable } from './environment';
+
+export function getAPIKey() {
+  try {
+    return getEnvVariable('XATA_API_KEY') ?? XATA_API_KEY;
+  } catch (err) {
+    return undefined;
+  }
+}
